Moved code from App.jsx to Home.jsx
App.jsx is now cleaner, and Home.jsx holds more of what it was intended to do.
